// src/app/_app-schema/types/parsed-project.ts
export type ParsedProject = {
  id: string;
  name: string;
  distance: string;
  elevation: string;
  time: string;
  companionship: string;
  islands?: string;
  munros?: string;
  munroTops?: string;
  corbetts?: string;
  corbettTops?: string;
  grahams?: string;
  subTwos?: string;
  donalds?: string;
};